fix types

+++ Jaberah/Models/JaberahModels/FollowStudentInMonthRow.cs
@@ -8,8 +8,8 @@
         public int WithFriendId { get; set; }
         public WithTeacherFriend WithTeacher { get; set; }
         public WithTeacherFriend WithFriend { get; set; }
-        public float Attendance { get; set; }
-        public float Behavior { get; set; }
+        public byte Attendance { get; set; }
+        public byte Behavior { get; set; }
         public int FollowStudentInMonthId { get; set; }
         public FollowStudentInMonth FollowStudentInMonth { get; set; }
     }
